backend-clean/requirements.txt
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
